+++ src/main/java/com/wccg/well_c_code_git_backend/domain/accesstoken/service/AccessTokenService.java
@@ -1,6 +1,9 @@
-package com.wccg.well_c_code_git_backend.domain.accesstoken;
+package com.wccg.well_c_code_git_backend.domain.accesstoken.service;
 
-import com.wccg.well_c_code_git_backend.domain.user.User;
+import com.wccg.well_c_code_git_backend.domain.accesstoken.repository.AccessTokenRepository;
+import com.wccg.well_c_code_git_backend.domain.accesstoken.dto.AccessTokenSaveRequest;
+import com.wccg.well_c_code_git_backend.domain.accesstoken.model.AccessToken;
+import com.wccg.well_c_code_git_backend.domain.user.model.User;
 import com.wccg.well_c_code_git_backend.global.exception.exceptions.AccessTokenNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
