+++ src/main/java/com/wccg/well_c_code_git_backend/domain/user/dto/UserSaveRequest.java
@@ -1,5 +1,6 @@
-package com.wccg.well_c_code_git_backend.domain.user;
+package com.wccg.well_c_code_git_backend.domain.user.dto;
 
+import com.wccg.well_c_code_git_backend.domain.user.model.UserRole;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
