+++ src/main/java/com/wccg/well_c_code_git_backend/domain/user/model/UserRole.java
@@ -1,4 +1,4 @@
-package com.wccg.well_c_code_git_backend.domain.user;
+package com.wccg.well_c_code_git_backend.domain.user.model;
 
 public enum UserRole {
     USER("사용자"),
