+++ src/main/java/com/wccg/well_c_code_git_backend/domain/user/repository/UserRepository.java
@@ -1,5 +1,6 @@
-package com.wccg.well_c_code_git_backend.domain.user;
+package com.wccg.well_c_code_git_backend.domain.user.repository;
 
+import com.wccg.well_c_code_git_backend.domain.user.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
