+++ src/main/java/com/wccg/well_c_code_git_backend/domain/user/service/UserService.java
@@ -1,5 +1,8 @@
-package com.wccg.well_c_code_git_backend.domain.user;
+package com.wccg.well_c_code_git_backend.domain.user.service;
 
+import com.wccg.well_c_code_git_backend.domain.user.model.User;
+import com.wccg.well_c_code_git_backend.domain.user.dto.UserSaveRequest;
+import com.wccg.well_c_code_git_backend.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
