+++ src/main/java/com/wccg/well_c_code_git_backend/domain/wccgrepository/controller/WccgRepositoryController.java
@@ -1,11 +1,14 @@
-package com.wccg.well_c_code_git_backend.domain.wccgrepository;
+package com.wccg.well_c_code_git_backend.domain.wccgrepository.controller;
 
-import com.wccg.well_c_code_git_backend.domain.user.User;
-import com.wccg.well_c_code_git_backend.domain.user.UserService;
+import com.wccg.well_c_code_git_backend.domain.user.model.User;
+import com.wccg.well_c_code_git_backend.domain.user.service.UserService;
+import com.wccg.well_c_code_git_backend.domain.wccgrepository.model.RepositorySortType;
+import com.wccg.well_c_code_git_backend.domain.wccgrepository.model.SortDirection;
+import com.wccg.well_c_code_git_backend.domain.wccgrepository.service.WccgRepositoryService;
 import com.wccg.well_c_code_git_backend.domain.wccgrepository.dto.GetRepositoriesResponse;
 import com.wccg.well_c_code_git_backend.domain.wccgrepository.dto.ServiceGetRepositoriesResponse;
-import com.wccg.well_c_code_git_backend.domain.wccgrepository.dto.WccgRepositoryServiceSyncResponse;
-import com.wccg.well_c_code_git_backend.domain.wccgrepository.dto.WccgRepositorySyncResponse;
+import com.wccg.well_c_code_git_backend.domain.wccgrepository.dto.ServiceSyncResponse;
+import com.wccg.well_c_code_git_backend.domain.wccgrepository.dto.SyncResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -26,10 +29,10 @@ public class WccgRepositoryController {
 
     @PreAuthorize("hasRole('USER')")
     @PostMapping("/sync")
-    public ResponseEntity<WccgRepositorySyncResponse> sync(@AuthenticationPrincipal User user) {
-        WccgRepositoryServiceSyncResponse ServiceResponse = wccgRepositoryService.syncRepositoryFrom(user);
+    public ResponseEntity<SyncResponse> sync(@AuthenticationPrincipal User user) {
+        ServiceSyncResponse ServiceResponse = wccgRepositoryService.syncRepositoryFrom(user);
 
-        WccgRepositorySyncResponse ControllerResponse = new WccgRepositorySyncResponse(ServiceResponse.getRepositoryCount());
+        SyncResponse ControllerResponse = new SyncResponse(ServiceResponse.getRepositoryCount());
 
         return ResponseEntity
                 .status(HttpStatus.OK)
