+++ src/main/java/com/wccg/well_c_code_git_backend/domain/wccgrepository/dto/ServiceSyncResponse.java
@@ -7,6 +7,6 @@ import lombok.ToString;
 @Getter
 @ToString
 @RequiredArgsConstructor
-public class WccgRepositoryServiceSyncResponse {
+public class ServiceSyncResponse {
     private final int repositoryCount;
 }
