+++ src/main/java/com/wccg/well_c_code_git_backend/domain/wccgrepository/dto/SyncResponse.java
@@ -7,6 +7,6 @@ import lombok.ToString;
 @Getter
 @ToString
 @RequiredArgsConstructor
-public class WccgRepositorySyncResponse {
+public class SyncResponse {
     private final int repositoryCount;
 }
