+++ src/main/java/com/wccg/well_c_code_git_backend/domain/wccgrepository/repository/WccgRepositoryRepository.java
@@ -1,5 +1,6 @@
-package com.wccg.well_c_code_git_backend.domain.wccgrepository;
+package com.wccg.well_c_code_git_backend.domain.wccgrepository.repository;
 
+import com.wccg.well_c_code_git_backend.domain.wccgrepository.model.WccgRepository;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.repository.JpaRepository;
