+++ src/main/java/com/wccg/well_c_code_git_backend/global/security/oauth/dto/LoginResponse.java
@@ -1,6 +1,6 @@
 package com.wccg.well_c_code_git_backend.global.security.oauth.dto;
 
-import com.wccg.well_c_code_git_backend.domain.user.User;
+import com.wccg.well_c_code_git_backend.domain.user.model.User;
 import lombok.Getter;
 
 @Getter
